Autorestart after pecl extension creation
Scenario: After your first puppet run you notice that the phing install command has returned 255 exit code instead of 0, so puppet marks it as failed, and because other pecl modules depend on it, it skips installing them. On your next run everything works fine, phing and the modules install but the httpd service does not get restarted.

The fix aims to notify the service in the above scenario.